Create RedactionConfigBlobstore
Summary:
This diff adds a class RedactionConfigBlobstore, which can be built using `RepoFactory`.

It is a prefix blobstore (with prefix "redactionconfig") that's created independent to all other repo blobstores. And it's where we'll store the blobs related to redaction key lists.

I created this because AFAICT using RepoFactory we can only build RepoBlobstore, which are partitioned by repo, which we don't want for redaction as it is generic for all repos.

It's not used in this diff, will be used in the future, I split it mostly to make it easier to review.

Reviewed By: markbt

Differential Revision: D29033599

fbshipit-source-id: 5c44a73e2097c0de3abad038f45166f42a14a70b