blame_v2: fix parent index determination in mononoke_admin
Summary:
The parent index for blame ranges is the index of the first *changeset* parent
that contains the file, not the index of the unode parent, which is what is
currently used for non-copy parents.  In fact, unode parents are guaranteed to
exist, so currently this will always be 0.

Change the calculation to work out which changeset parent contains the file.

Differential Revision: D29453805

fbshipit-source-id: 98950090a7d87e4ece22d8047dc3794ac52ec4a2