unodes: fix find_unode_renames for files copied multiple times
Summary:
`find_unode_renames` uses a hashmap for renames keyed on the source path.  This
means when files are copied multiple times, only the last copy is kept.

Implement a new version, which takes multiple renames into account.  Prepare
for blame V2 by including the parent index and source path in the generated
rename source.

The existing implementation is retained so that we don't change how blame V1 is
computed.

Reviewed By: StanislavGlebik

Differential Revision: D29454872

fbshipit-source-id: e6de5cad3f00ac8413d25c385d42619c8cb02a31