Initialise redaction config using configerator
Summary:
This diff joins all previous diffs, by making redaction config be read from configerator by default, when created from RepoFactory.

The logic for fetching from XDB still exists, and can be used if a tunable is enabled, though this needs a service restart. I plan to leave it there as a possible fallback until we successfully add some real-world thing to the new redaction config and it works.

Reviewed By: StanislavGlebik

Differential Revision: D29033596

fbshipit-source-id: 4c5e97f542457dc25cf234d31182c3106173585d