mononoke/blobstore: new `is_present` semantics via enum
Summary:
This diff introduces new `is_present` semantics that will allow to move the decision logic on the complex multiplex `is_present` result from the multiplex to the callers.

In the current API `is_present` call returns `Result<bool>` and in case it couldn't determine whether the key exists or not, it checks the sync-queue in case the key was written recently and then might check the multiplex stores again, then fails if still unsure. This brings unnecessary complications and makes the multiplex blobstore less reliable.
More details in: https://fb.quip.com/wOCeAhGx6Oa1

This change allows us to get rid of the queue and second store lookups and move the decision-making to the callers.

*This diff shouldn't change the behaviour, but to replace bool value with an enum and add the conversions where needed.*

Reviewed By: StanislavGlebik

Differential Revision: D29377462

fbshipit-source-id: 4b70f772d2ed70d9fffda015ba06c3f16bf1475d