pull: filter out existing heads to pull
Summary:
If heads exist in the repo, there is no need to pull them.

Practically we configured selectivepull to include master and stable.
While the master head is excluded by the pull fast path, the stable bookmark
previously triggers the heavyweight pull. This diff makes it that we can
skip the heavyweight pull and avoid other issues like devel-warn importing
empty changegroup.

Reviewed By: DurhamG

Differential Revision: D29525476

fbshipit-source-id: 9d1ff28d1194cac22ae66e669a5bd9dbe3f750c4