Extract periodic reloader to common class and use it in skiplist
Summary:
I've seen periodic reloading of stuff in at least 3 places in mononoke (2 of which I added, skiplists and redaction config, and also on segmented changelog, there might be more).

This stack extracts that logic to a common place, so we don't need to reinvent that logic all the time, and it's easier to do it the next time.

Reviewed By: mitrandir77

Differential Revision: D29520651

fbshipit-source-id: 59820c03f168cb25e2c6345e36746121451f34e2