checkout: fix issue with symlink=>directory replacement
Summary:
We had an issue in native checkout when update needed to remove a symlink and then create a directory with same files instead of symlink.
This used to fail, because update has a plan to write to new files, but the files has already 'existed' as part of a symlink, so unknown files check were failing.

This code makes sure that when listing untracked files we do not go inside symlinks, and treat audit errors from VFS as if file did not exist

Reviewed By: DurhamG

Differential Revision: D29567562

fbshipit-source-id: 1b6751cc00c3c628e2cab8c081540dba200209fa