Implement uploading of hg changesets
Summary:
implement uploading of hg changesets

For now, reuse the upload code path from unbundle but calling it with empty filenodes and manifests.

Those are used for parents validation but this is not needed for us because we load trees and filenodes and their parents to construct the bonsai cs.

We might want to rewrite it to a cleaner code later and separate from unbundle but for now reusing the fucntion is the easiest way because we know the implementation is correct and also has logging.

Reviewed By: markbt

Differential Revision: D29644798

fbshipit-source-id: 27217d3061ab8d9712417facdbfbbc7e3aebfc5b