blame: extract function for fetching content
Summary:
Extract `fetch_content_for_blame` to a separate module so we can re-use it in
blame V2.

The method previously returned nested `Result`s, which can be confusing as in
most contexts, the blame being rejected is not actually an error.  Switch to an
explicit enum to make it clearer what the inner result represents.

Reviewed By: yancouto

Differential Revision: D29462095

fbshipit-source-id: 52ffcb4173a3b36f4b6cdafe4f42a4cafd993f49