revisionstore: unify "store type" (local vs. shared/cache) enums
Summary:
We have a variety of enums in revisionstore that all serve the same purpose. With this change, I've consolidated them to a single type.

Not 100% sure if `StoreType` is the best name, in scmstore I named it `LocalStoreType` to be more clear - open to suggestions. I also decided to keep the variants named `Local` and `Shared` instead of adopting the `Local` and `Cache` terminology I used in scmstore - I'd rather not change that unless we decide to change the terminology everywhere to avoid confusion.

Reviewed By: kulshrax

Differential Revision: D29598025

fbshipit-source-id: 76d5a02230a8c1e5327cc5d90bbcae70049f251d