edenapi: introduce file attributes support
Summary:
Add support for optional file attributes to EdenApi, with `content` there as a placeholder.

Modifies the `FileRequest` type, adding a vec of `FileSpec`, which allows the client to specify desired attributes per-key. The existing `keys` field will be treated as a request for the content attribute and may be used in combination with the new per-key attributes.

Reviewed By: DurhamG

Differential Revision: D29634709

fbshipit-source-id: 6571837f87d1635e8529490e10dbe4ba054b7348