memcachestore: remove direct usage of MemcacheStore
Summary:
Previously we could add a MemcacheStore to the store hierarchy in two
ways: directly as MemcacheStore and through the MemcacheHgIdData/HistoryStore
wrappers.

In a future diff we'll want to track how long since the store was created and
not call the inner MemcacheStore in certain situations. To do this, we want all
accesses to go through the MemcacheHgId*Store abstraction.

Reviewed By: quark-zju

Differential Revision: D29627226

fbshipit-source-id: 9e979281fbb2eec123577d99a8879bcf80578136