add auto fsck path for snapshot tests
Summary: Currently the tests manually run fsck command (Python) on the snapshot. This makes the change so that the tests will not only test the manual fsck command but also test the auto fsck (C++) so that we cover both Eden CLI fsck (Python) and Eden mount auto fsck (C++).

Reviewed By: chadaustin

Differential Revision: D29690188

fbshipit-source-id: 593db1db587c3294aad1314ea8da1d8e778df8ee