overwrite headers with config and verbose mode
Summary:
It's useful to be able to overwrite or inject http headers without changing client's source code. I needed that when I was proxying client requests through ncat running locally on my machine - I had to set Host header.
Moreover, if it's part of the config, we'll be able to distribute it quickly with dynamicconfig to do some things.

Also, we have http.verbose that's being used by rust http client but it's not respected by mononokepeer in python. I'm adding that here as well. Not full blown verbose mode because it's not necessary - we're only upgrading to websocket.

Reviewed By: farnz

Differential Revision: D29792386

fbshipit-source-id: 51b7f3cf07a870636ac6aa126f0efb45e979ef30