add support for issue CAT commit cloud tokens on corp
Summary:
Add support for issue CAT commit cloud tokens on corp Macs.

This is needed to support enabling Scm Daemon for all mac users **automatically**.

The tool `corp_clicat` is probably not yet deployed everywhere but it will work once deployed.

The current manual way will remain working as well.

The current way is manual:

* obtaining OAuth token from https://our.intern.facebook.com/intern/oauth/184975892288525

* running `hg cloud auth -t <token>` (the command stores the token into Mac keychain, Scm Daemon fetches from it).

The problem with the manual way is that users are not aware about scm daemon and the features it provides. The manual way worked fine when the tokens were also used for hg commit cloud extension as well but now they are not used.

Reviewed By: markbt

Differential Revision: D29791737

fbshipit-source-id: 2391a196167e1f17d53a3231dbf58f7cb2bcd39a