dag: add fail points on opening namedag
Summary:
It turns out that the namedag was opened multiple times. Add a fail point to
help figure out the callsite.

The `fail` crate allows something like:

  FAILPOINTS="dag-namedag-open=1*sleep(1)->return"
  FAILPOINTS="dag-namedag-open=1*sleep(1)->panic"

Meaning that the first open causes 1ms sleep, and the second
causes an error (turns into a Python backtrace), or a panic (turns into a Rust
backtrace with RUST_BACKTRACE=1).

Reviewed By: andll

Differential Revision: D29888937

fbshipit-source-id: b1644d7196f68262523ab9a5fc4fb110a4cc0062