Add bonsai upload to eden api py ext
Summary:
## High level goal

This stack aims to add a way to upload commits directly using the bonsai format via edenapi, instead of using the hg format and converting on server.

The reason this is necessary is that snapshots will be uploaded on bonsai format directly, as hg format doesn't support them. So this is a stepping stone to do that, first being implemented on commit cloud upload, as that code already uses eden api, and later will be used by the snapshotting commands.

## This diff

This diff adds the new endpoint to the API that can be acessed via python code, and in turn calls the api implemented on the last diff.

Reviewed By: liubov-dmitrieva

Differential Revision: D29849962

fbshipit-source-id: 5a2a674aef1edd3b0d95cb2b45b02ef9c20aca48