Add upload_fileblobs function
Summary:
## High level goal

This stack aims to add a way to upload commits directly using the bonsai format via edenapi, instead of using the hg format and converting on server.

The reason this is necessary is that snapshots will be uploaded on bonsai format directly, as hg format doesn't support them. So this is a stepping stone to do that, first being implemented on commit cloud upload, as that code already uses eden api, and later will be used by the snapshotting commands.

## This diff

Uploads file contents without uploading filenodes. Will be used for uploading a
commit in bonsair format instead of mercurial format via EdenAPI.

This is basically the same method as before D29549091 (https://github.com/facebookexperimental/eden/commit/d327996144b9f973720a6b61db79a66442956e99).

Reviewed By: liubov-dmitrieva

Differential Revision: D29799484

fbshipit-source-id: 136c058ebcd814f39c5b903f5d8bfef7ff6005dc