Add upload bonsai changeset endpoint
Summary:
## High level goal

This stack aims to add a way to upload commits directly using the bonsai format via edenapi, instead of using the hg format and converting on server.

The reason this is necessary is that snapshots will be uploaded on bonsai format directly, as hg format doesn't support them. So this is a stepping stone to do that, first being implemented on commit cloud upload, as that code already uses eden api, and later will be used by the snapshotting commands.

## This diff

This diff creates an endpoint on eden api which uploads a commit using the bonsai format.

It also adds all the necessary types to represent a bonsai commit (basically the same as hg commit, but no manifests, and a bit more detail on how each file changed) via the wire, and related boilerplate.

Reviewed By: liubov-dmitrieva

Differential Revision: D29849963

fbshipit-source-id: 2ff44d53874449ae4373a0135a60ead40c541309