Add bonsai upload to EdenApi trait
Summary:
## High level goal

This stack aims to add a way to upload commits directly using the bonsai format via edenapi, instead of using the hg format and converting on server.

The reason this is necessary is that snapshots will be uploaded on bonsai format directly, as hg format doesn't support them. So this is a stepping stone to do that, first being implemented on commit cloud upload, as that code already uses eden api, and later will be used by the snapshotting commands.

## This diff

This diff adds a trait method to call the endpoint added on D29849963.

It's mostly boilerplate, calling the correct endpoint, converting types from/to wire.

Reviewed By: liubov-dmitrieva

Differential Revision: D29849965

fbshipit-source-id: 4a821e965fe4319fddd8e13b13ed4de5b7f86e93