amend: fix file mixup during "--to" rebase
Summary:
When rebasing during "amend --to", we were mixing up file contents when there was more than a single merged file in a three way merge. The cause was a lambda within a loop closing over a loop variable.

Also suppress the "merging some/file.c" message if the result of the three way merge is identical to the "local" version of the file. It is confusing to see unexpected "merging" messages.

Reviewed By: DurhamG

Differential Revision: D29940097

fbshipit-source-id: 5a4c19279c14209268359939fbf91f164c791b2e