fix debughiddencommit command for LFS
Summary:
fix `debughiddencommit` command for LFS

There was an issue with the current implementation. It was hidden too early.

backup.backup logic triggers lfs upload but because lfs extension is not enabled, it goes to

remotefilelog.uploadblobs. This function contains the following code:

```
def uploadblobs(repo, nodes):
    toupload = []
    for ctx in repo.set("%ln - public()", nodes):
        for f in ctx.files():
            if f not in ctx:
                continue

            fctx = ctx[f]
            toupload.append((fctx.path(), fctx.filenode()))
    repo.fileslog.contentstore.upload(toupload)
```

The revset doesn't work for commits.

Reviewed By: markbt

Differential Revision: D29960636

fbshipit-source-id: 8c746c5b0484678e9e988105e980381a0172b38c