nativecheckout: detect unchanged sparse profile
Summary: If the "update" destination sparse profile is identical to the current sparse profile, we now pass "None" to checkoutplan, allowing it to avoid slow call to with_sparse_profile_change. Updating across a small number of commits still took a while due to the unnecessary futzing with sparse profiles.

Reviewed By: andll

Differential Revision: D29949015

fbshipit-source-id: d6fc2ceb7776cd2f2ee42f3dc5b8358f770a7947