Update backup state in `hg cloud upload` command
Summary:
Update backup state in `hg cloud upload` command

The backup state is used by `hg sl`, so it would be nice to keep it up-to-date after `hg cloud upload` command, similar to old `hg cloud backup`.

Also, we should add heads what we filtered in order to update the backup state correctly.

So, it will now returned list of uploaded heads as nodes (including filtered) and list of failed commits as nodes (not only heads).

Reviewed By: markbt

Differential Revision: D29878296

fbshipit-source-id: 5848e9f86175fbdc56db123cf7ba0d5fc51273b0