configparser: drop thrift-types dep
Summary:
The thrift-types dep is quite heavyweight. It depends on fbthrift, and then
tokio-shim, and needs complex code gen logic. It is especially annoying in open
source builds (ex on non-fbcode Linux) since building the fbthrift with
compatible dependencies is quite slow and error prone.

The configparser is used by backingstore, which is used by edenfs. Decoupling
from thrift-types (and fbthrift, tokio-shim, etc.) avoids a lot of headache for edenfs
build.

The actual logic we need here is only JSON serialization for some simple
structs. There is no need to use Thrift serialization at all.

Let's just check-in the lightweight Rust structs that support serde JSON
serialization.

Ideally, the Thrift compiler could generate code that is serde struct only
without the fbthrift stuff. However, it does not support that for now.

A `build.rs` script is added to prevent the code from being out of sync
with the Thrift source. It seems fine given that the Thrift source is
rarely changed.

Reviewed By: DurhamG

Differential Revision: D30008264

fbshipit-source-id: 1822c3f0d47445a9e59d26575a567cb51b51cec0